requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
fastapi==0.66.0
greenlet==1.1.0
h11==0.12.0
Jinja2==3.0.1
MarkupSafe==2.0.1
pydantic==1.8.2
PyJWT==2.1.0
python-multipart==0.0.5
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.21
SQLAlchemy-Utils==0.37.8
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
mysqlclient==2.1.0